Fix : " 주술사, 도사 시전 속도 타입, 값 수정 "
GameServer
 - PlayerCharacterStatus
  - 주술사, 도사 시전속도 값을 int16 -> float으로 수정
 - Data, DataManager
  - 수정된 데이터 값에 맞게 파싱
 - GameObjectInfo
  - st_StatInfo 수정
   - MagicHitRate int16 -> float 로 수정

+++ GameServer/Data.h
@@ -117,7 +117,7 @@ struct st_ObjectStatusData
 	int32 MaxMeleeAttackDamage;
 	int16 MeleeAttackHitRate;
 	int16 MagicDamage;
-	int16 MagicHitRate;
+	float MagicHitRate;
 	int32 Defence;
 	int16 EvasionRate;
 	int16 MeleeCriticalPoint;
