+++ GameServer/GameObjectInfo.h
@@ -1701,7 +1701,7 @@ struct st_StatInfo
 	int32 MaxMeleeAttackDamage;
 	int16 MeleeAttackHitRate;
 	int16 MagicDamage;
-	int16 MagicHitRate;
+	float MagicHitRate;
 	int32 Defence;	
 	int16 EvasionRate;
 	int16 MeleeCriticalPoint;
